fitnes-club-main/requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
pre-commit
django-filter
Pillow
djangorestframework-simplejwt
dj-rest-auth
django-ckeditor